requirements.txt
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.3.0
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
PySSTV==0.5.7
pytesseract==0.3.13
python-bidi==0.6.6
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
sympy==1.13.1
tifffile==2025.3.13
tomli_w==1.2.0
torch==2.6.0
torchvision==0.21.0
triton==3.2.0
typing_extensions==4.12.2
wcwidth==0.2.13
